[6921] Reorganize some of the WIP Redux files for sync tab, per PR feedback. Move action and state into separate files.
